adding CAN and websockets
websockets_demo is the actual websockets code and can_receive_websockets_new.ino is an arduino version of the platformIO code we have for the receive node (second node) with joystick and rpm currently. so websockets_demo should be run on the front daq node, the one currently with the temp sensor.